Big update: -Equipment: Fixed some bugs in there as well as changes to add Variants.
-Hooks: We have 6 new hooks coming in to be added by Pyro. Two of those are added at the top until he actually gets them into the updater. The other 4 need to be added to constants in Hooks.scar but currently reside in their raw form in Interfaces in AbsolutePos.

-Interfaces: Big changes. Added Absolute coords! These will be automatically calculated when GetInterfaceComponent or child component is called. Also added actions to the GetChildComponent. Other minor bug fixes/changes.

-Inventory: Minor changes/bugfixes. Nothin too major here.

-Players: Minor changes/bug fixes. Made sure Fighting works properly. Probably need to change some of the records around to make names for certain things more universal. Will take another look at players sometime this week.

-Text: Minor changes.

ADDED: Bank.simba - Ultra makeover to the bank system. This include is much more stable and works like a charm. A lot of things incorporated to use the Absolute positioning in interfaces. Withdrawing rewritten to use scrolling to find items. Great changes, and should really benefit all banking scripts everywhere!

ADDED: Depositbox.simba: Added 2 functions in there for now that Echo had in the previous reflection include. This will house more functions for it later. Put it in the misc folder because I don't feel it needs to be in "core".

ADDED: PathMaker.simba: Useful little tool for recording tiles. Fixed it up to work with this version of the include.

Signed-off-by: Drags111 <jdeanwaite@gmail.com>